requirements.txt
python-dotenv==1.1.0
PyMuPDF==1.26.0
PyPDF2==3.0.1
groq==0.26.0
mysql-connector==2.2.9
pandas==2.3.0
sentence-transformers==4.1.0
faiss-cpu==1.11.0
numpy==2.3.0
scikit-learn==1.7.0
Flask==3.1.1

